// front/projects/shared-components/src/lib/models/table-data-changes.ts
/**
 * Модель изменений данных в таблице.
 * Хранит идентификаторы добавленных, изменённых и удалённых элементов.
 */
export interface TableDataChanges {
    /** Массив ID добавленных элементов */
    added: string[];
    /** Массив ID изменённых элементов */
    modified: string[];
    /** Массив ID удалённых элементов */
    deleted: string[];
}